Complete test for 'grouper' function factory
