+++ tests/fixtures/iterating.py
@@ -1,16 +1,7 @@
-from typing import (Any,
-                    Hashable)
-
 import pytest
 
-from lz.hints import Map
 from tests import strategies
 from tests.utils import find
-
-
-@pytest.fixture(scope='function')
-def grouper_key() -> Map[Any, Hashable]:
-    return find(strategies.groupers_keys)
 
 
 @pytest.fixture(scope='function')
