+++ tests/iterating_tests/grouper_tests/test_definition.py
@@ -2,15 +2,18 @@ from typing import (Any,
                     Hashable,
                     Iterable)
 
+from hypothesis import given
+
 from lz.hints import Map
 from lz.iterating import grouper
 from lz.replication import duplicate
+from . import strategies
 
 
-def test_basic(hashables_iterable: Iterable[Hashable],
-               grouper_key: Map[Any, Hashable]) -> None:
-    original, target = duplicate(hashables_iterable)
-    group_by = grouper(grouper_key)
+@given(strategies.hashables_iterables, strategies.groupers_keys)
+def test_basic(iterable: Iterable[Hashable], key: Map[Any, Hashable]) -> None:
+    original, target = duplicate(iterable)
+    group_by = grouper(key)
 
     result = group_by(target)
     result_list = list(result)
@@ -18,4 +21,4 @@ def test_basic(hashables_iterable: Iterable[Hashable],
     assert all(element in group
                for element in original
                for key, group in result_list
-               if grouper_key(element) == key)
+               if key(element) == key)
