+++ tests/iterating_tests/grouper_tests/test_pickle.py
@@ -2,17 +2,21 @@ from typing import (Any,
                     Hashable,
                     Iterable)
 
+from hypothesis import given
+
 from lz.hints import Map
 from lz.iterating import grouper
 from lz.replication import duplicate
 from tests.utils import (are_iterables_similar,
                          round_trip_pickle)
+from . import strategies
 
 
-def test_round_trip(plain_hashables_iterable: Iterable[Hashable],
-                    grouper_key: Map[Any, Hashable]) -> None:
-    original, target = duplicate(plain_hashables_iterable)
-    group_by = grouper(grouper_key)
+@given(strategies.plain_hashables_iterables, strategies.groupers_keys)
+def test_round_trip(iterable: Iterable[Hashable],
+                    key: Map[Any, Hashable]) -> None:
+    original, target = duplicate(iterable)
+    group_by = grouper(key)
 
     result = round_trip_pickle(group_by)
