+++ tests/strategies/__init__.py
@@ -26,8 +26,7 @@ from .functional import (
     transparent_functions_calls,
     true_predicates,
     various_suitable_maps)
-from .iterating import (groupers_keys,
-                        non_negative_indices,
+from .iterating import (non_negative_indices,
                         non_negative_slices,
                         positive_indices)
 from .literals import empty
